refactor all neurons to use check for registration
