+++ storage/shared/checks.py
@@ -117,3 +117,17 @@ def _get_redis_password(redis_conf_path):
         assert False, f"An error occurred: {e}"
 
     return None
+
+
+def check_registration(subtensor, wallet, netuid):
+    if not subtensor.is_hotkey_registered(
+        netuid=netuid,
+        hotkey_ss58=wallet.hotkey.ss58_address,
+    ):
+        bt.logging.error(
+            f"Wallet: {wallet} is not registered on netuid {netuid}"
+            f"Please register the hotkey using `btcli subnets register` before trying again"
+        )
+        exit()
+
+    pass
