+++ storage/shared/utils.py
@@ -153,17 +153,3 @@ def get_redis_password(
         exit(1)
 
     return redis_password
-
-
-def check_registration(subtensor, wallet, netuid):
-    if not subtensor.is_hotkey_registered(
-        netuid=netuid,
-        hotkey_ss58=wallet.hotkey.ss58_address,
-    ):
-        bt.logging.error(
-            f"Wallet: {wallet} is not registered on netuid {netuid}"
-            f"Please register the hotkey using `btcli subnets register` before trying again"
-        )
-        exit()
-
-    pass
